requirements.txt
python-dotenv
deeplake
csv
